Move types folder into visualization folder
